update 3d test.py
